changed post.js to brewery.js, split into 2 models... user.js and brewery.js
