--- models/post.js
@@ -1,60 +1,0 @@
-// Requiring bcrypt for password hashing. Using the bcryptjs version as the regular bcrypt module sometimes causes errors on Windows machines
-const bcrypt = require("bcryptjs");
-// Creating our User and Beer model
-
-module.exports = function (sequelize, DataTypes) {
-  var Beer = sequelize.define("Beer", {
-    //   EXAMPLE TO FOLLOW ONCE TABLE STRUCTURE IS DECIDED UPON
-
-    id: {
-       type: DataTypes.INTEGER,
-       allowNull: false,
-       primaryKey: true,
-       autoIncrement: true
-  
-    },
-    brewery_name: {
-       type: DataTypes.TEXT,
-       allowNull: false,
-       validate: {
-         len: [5],
-       },
-     },
-     tried: {
-       type: DataTypes.BOOLEAN,
-       defaultValue: false,
-     }
-  });
- 
-  const User = sequelize.define("User", {
-    // The email cannot be null, and must be a proper email before creation
-    email: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      unique: true,
-      validate: {
-        isEmail: true
-      }
-    },
-    // The password cannot be null
-    password: {
-      type: DataTypes.STRING,
-      allowNull: false
-    }
-  });
- //  Creating a custom method for our User model. This will check if an unhashed password entered by the user can be compared to the hashed password stored in our database
-   User.prototype.validPassword = function(password) {
-   return bcrypt.compareSync(password, this.password);
-   };
-  // Hooks are automatic methods that run during various phases of the User Model lifecycle
-  // In this case, before a User is created, we will automatically hash their password
-   User.addHook("beforeCreate", user => {
-     user.password = bcrypt.hashSync(
-       user.password,
-       bcrypt.genSaltSync(10),
-       null
-     );
-   });
-  
-  return Beer, User;
-};
